VBench-2.0/requirement.txt
accelerate==1.2.1
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
albucore==0.0.19
albumentations==1.4.19
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
args==0.1.0
async-timeout==5.0.1
attrs==24.3.0
av==14.0.1
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
blis==1.2.0
boto3==1.26.90
botocore
braceexpand==0.1.7
catalogue==2.0.10
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
clint==0.5.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpickle==3.1.1
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
datasets==3.2.0
decorator==5.1.1
decord==0.6.0
deep-sort-realtime==1.3.2
deepspeed==0.12.6
defusedxml==0.7.1
diffusers==0.32.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
easydict==1.13
einops==0.8.0
einops-exts==0.0.4
exceptiongroup==1.2.2
fairscale==0.4.13
fastapi==0.115.6
filelock==3.13.1
filterpy==1.4.5
fire==0.4.0
flowlib==0.6.6.8
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.2.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
h5py==3.12.1
hjson==3.1.0
hpsv2==1.2.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.5.0
iniconfig==2.0.0
iopath==0.1.10
ipdb==0.13.13
ipython==8.32.0
Jinja2==3.1.3
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.8
kornia==0.5.6
kornia_rs==0.1.8
langcodes==3.5.0
lazy_loader==0.4
-e git+https://github.com/LLaVA-VL/LLaVA-NeXT@79ef45a6d8b89b92d7a8525f077c3a3a9894a87d#egg=llava
logger==1.4
lvis==0.5.3
MarkupSafe==2.1.5
matplotlib==3.10.0
mmdet==3.0.0
mmengine==0.10.5
mmyolo==0.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
networkx==3.2.1
nltk==3.9.1
numpy==1.26.3
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.16.3
open_clip_torch==2.30.0
openai==1.59.5
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openmim==0.3.9
packaging==24.2
pandas==2.2.3
peft==0.14.0
pillow==10.2.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==3.1.1
preshed==3.0.9
prettytable==3.14.0
propcache==0.2.1
protobuf==3.20.3
py-cpuinfo==9.0.0
pyarrow==18.1.0
pycocoevalcap==1.2
pycocotools==2.0.8
pydantic==2.10.4
pydantic_core==2.27.2
pyecharts==2.0.8
pynvml==12.0.0
pyparsing==3.2.1
PySocks==1.7.1
pytest==7.2.0
pytest-split==0.8.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
qudida==0.0.4
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
scenedetect==0.6.5.2
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.15.0
seaborn==0.13.2
selenium==4.29.0
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
shapely==2.0.7
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
snapshot-selenium==0.0.2
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
srsly==2.5.1
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
tabulate==0.9.0
termcolor==2.5.0
terminaltables==3.1.10
thinc
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2025.1.10
tiktoken==0.8.0
timesformer-pytorch==0.4.1
timm==0.4.12
tokenizers==0.21.1
tomli==2.2.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.68
unicorn==2.1.1
urllib3==1.26.15
wandb==0.19.2
wasabi==1.1.3
wcwidth==0.2.13
webdataset==0.2.100
websocket==0.2.1
websocket-client==1.8.0
wordcloud==1.9.4
xxhash==3.5.0
yacs==0.1.8
yapf==0.43.0
yarl==1.18.3
-e git+https://github.com/AILab-CVC/YOLO-World.git@4826695a84ce370dad7fe4cb555a6c7162f3751d#egg=yolo_world
yolov5
zipp==3.21.0
mediapipe
retina-face
retinaface-pytorch
supervision
--extra-index-url https://download.openmmlab.com/mmcv/dist/cu118/torch2.4/
mmcv==2.2.0